+++ src/main/java/com/cs4256/drinkmorewater/controllers/ReviewController.java
@@ -1,6 +1,8 @@
 package com.cs4256.drinkmorewater.controllers;
 
 import com.cs4256.drinkmorewater.controllers.utils.R;
+import com.cs4256.drinkmorewater.controllers.utils.UserType;
+import com.cs4256.drinkmorewater.enums.TypeEnum;
 import com.cs4256.drinkmorewater.models.Review;
 import com.cs4256.drinkmorewater.models.User;
 import com.cs4256.drinkmorewater.services.impl.ReviewServiceImpl;
@@ -9,18 +11,31 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/review")
+@RequestMapping("/{uid}/review")
 public class ReviewController {
 
     @Autowired
     private ReviewServiceImpl reviewService;
+    @Autowired
+    private UserServiceImpl userService;
+
+    private UserType userType;
+
+    @ModelAttribute
+    private void createUserType(@PathVariable Integer uid) {
+        userType = new UserType(uid, userService.getById(uid).getType());
+    }
 
     /**
      * return all element
      * @return
      */
     @GetMapping
+    // user, admin, rest
     public R getAll() {
+        if (userType.getTypeEnum().equals(TypeEnum.ORDERAPP)) {
+            return new R(false, "You do not have right");
+        }
         return new R(true, reviewService.list());
     }
 
@@ -29,8 +44,32 @@ public class ReviewController {
      * @return
      */
     @GetMapping("/{id}")
+    // admin, user, rest
     public R getById(@PathVariable Integer id) {
+        if (userType.getTypeEnum().equals(TypeEnum.ORDERAPP)) {
+            return new R(false, "You do not have right");
+        }
         return new R(true, reviewService.getById(id));
+    }
+
+
+    // admin user rest
+    // get by restaurant id -- found all
+    @GetMapping("/user/{id}")
+    public R getUserRestReviewByUserId(@PathVariable Integer id) {
+        if (userType.getTypeEnum().equals(TypeEnum.ORDERAPP)) {
+            return new R(false, "You do not have right");
+        }
+        return new R(true, reviewService.selectUserRestReviewByUserId(id));
+    }
+
+    // get by user id -- found all
+    @GetMapping("/rest/{id}")
+    public R getUserRestReviewByRestId(@PathVariable Integer id) {
+        if (userType.getTypeEnum().equals(TypeEnum.ORDERAPP)) {
+            return new R(false, "You do not have right");
+        }
+        return new R(true, reviewService.selectUserRestReviewByRestId(id));
     }
 
     /**
@@ -38,8 +77,15 @@ public class ReviewController {
      * @return
      */
     @PostMapping
+    // admin, user(id)
     public R insert(@RequestBody Review review) {
-        return new R(reviewService.save(review));
+        if (userType.getTypeEnum().equals(TypeEnum.ADMIN)) {
+            return new R(reviewService.save(review));
+        } else if (userType.getTypeEnum().equals(TypeEnum.CUSTOMER) &&
+        userType.getUid().equals(review.getUserId())) {
+            return new R(reviewService.save(review));
+        }
+        return new R(false, "You do not have right");
     }
 
     /**
@@ -48,8 +94,12 @@ public class ReviewController {
      * @return
      */
     @PutMapping
+    // admin
     public R updateById(@RequestBody Review review) {
-        return new R(reviewService.updateById(review));
+        if (userType.getTypeEnum().equals(TypeEnum.ADMIN)) {
+            return new R(reviewService.updateById(review));
+        }
+        return new R(false, "You do not have right");
     }
 
     /**
@@ -57,7 +107,16 @@ public class ReviewController {
      * @return
      */
     @DeleteMapping("/{id}")
+    // admin, user(id)
     public R deleteById(@PathVariable Integer id) {
-        return new R(reviewService.removeById(id));
+        if (userType.getTypeEnum().equals(TypeEnum.ADMIN)) {
+            return new R(reviewService.removeById(id));
+        } else if (userType.getTypeEnum().equals(TypeEnum.CUSTOMER) &&
+        userType.getUid().equals(reviewService.getById(id).getUserId())) {
+            return new R(reviewService.removeById(id));
+        }
+        else {
+            return new R(false, "You do not have right");
+        }
     }
 }
